Start Text Font, fixed rules and credits screen, changed start menu music
Changed the font of the text saying "press space to start" and moved it down so i can fit the game's logo eventually. Also changed the music of the title screen and menu (i want to use the prev one for the igloo/village) and i want to find another music for the battle. Also implemented the moving bg to the rules, credits, and powerup screens.